JS tests for travis

+++ run_tests.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
-wget https://raw.github.com/jquery/qunit/master/addons/phantomjs/runner.js
-python manage.py runserver &
-SERVER=$!
-sleep 3
-phantomjs runner.js http://localhost:8000/static/tests/setup.html
-SUCCESS=$?
-kill $SERVER
-rm runner.js
-if python manage.py test; then
+if [ "$1" == "js" ]; then
+    wget https://raw.github.com/jquery/qunit/master/addons/phantomjs/runner.js
+    python manage.py runserver &
+    sleep 3
+    SERVER=$!
+    phantomjs runner.js http://localhost:8000/static/tests/setup.html
+    SUCCESS=$?
+    kill $SERVER
+    rm runner.js
     exit $SUCCESS
+else
+    exec python manage.py test "$1"
 fi
-exit 1
